feat(api): package exports swaggerlint & types
